almost complete cdh5_oozie, added cdh5_hive

+++ cloudera_cdh_docker/scripts/0_startdns.sh
@@ -1,4 +1,13 @@
-docker stop dns
-docker rm dns
-docker run -d -t --name dns -h dns.example.com -v /etc/bind pti1/bind9:initialversion
+containerExists=`docker ps -a | grep "dns"`
+
+if [ "$containerExists" ];then
+
+	docker stop dns
+	#Remove the container
+	docker rm dns
+fi
+
+
+
+docker run -d -t --name dns -h dns.example.com -v /etc/bind pti1/bind9:secondversion
